week3/server/requirements.txt
mcp>=0.9.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
tenacity>=8.2.3